codeextraction/requirements.txt
pymupdf>=1.24.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
